CSS all the things.
All the things are now horizontal or vertical cells, and define their
user agent style sheet for defaulting.

+++ kramer/src/main/java/com/chiralbehaviors/layout/ColumnSet.java
@@ -66,7 +66,7 @@ public class ColumnSet {
         }
 
         // compression
-        double columnWidth = relax(justified / (double) count);
+        double columnWidth = relax(justified / count);
         firstColumn.setWidth(columnWidth);
         double compressed = relax(columnWidth - labelWidth);
         firstColumn.getFields()
